advanced_python/requirements.txt
Flask
Flask-JWT-Extended
oauthlib
Flask-RESTful
Flask-SQLAlchemy
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
requests
python-dotenv
flask-uploads
flask-migrate
psycopg2-binary==2.8.4
requests-oauthlib==1.1.0
flask-oauthlib ==0.9.0


